Rename and change operation from order api -> order service to order service -> inventory service
