+++ Models/Order.go
@@ -12,7 +12,7 @@ import (
 type Order struct {
 	ID        uuid.UUID `json:"id" swaggerignore:"true"`
 	ProductID string    `json:"product_id" validate:"required"`
-	Qty       int       `json:"qty" validate:"required"`
+	Qty       uint      `json:"qty" validate:"required"`
 	Status    string    `json:"status" swaggerignore:"true"`
 	CreatedAt time.Time `json:"created_at" swaggerignore:"true"`
 	UpdatedAt time.Time `json:"updated_at" swaggerignore:"true"`
