--- OrderConsumer/Core/Ports/Repositories/OrderRepository.go
@@ -1,7 +1,0 @@
-package ports_repositories
-
-import models "github.com/Prompiriya084/go-mq/Models"
-
-type OrderRepository interface {
-	repository[models.Order]
-}
