--- OrderConsumer/Core/Services/OrderService.go
@@ -1,9 +1,0 @@
-package services
-
-import models "github.com/Prompiriya084/go-mq/Models"
-
-type OrderService interface {
-	Create(order *models.Order) error
-	Update(order *models.Order) error
-	Delete(order *models.Order) error
-}
