+++ OrderService/Internal/Core/Utilities/Validator/Validator.go
@@ -1,0 +1,5 @@
+package utilities_validator
+
+type Validator interface {
+	ValidateStruct(s interface{}) error
+}
