--- Producer/Internal/Core/Ports/Repositories/ProductRepository.go
@@ -1,9 +1,0 @@
-package ports_repositories
-
-import (
-	models "github.com/Prompiriya084/go-mq/Models"
-)
-
-type ProductRepository interface {
-	repository[models.Product]
-}
