--- Producer/Internal/Core/Services/ProductServiceImpl.go
@@ -1,1 +1,0 @@
-package services
