--- Producer/Internal/Core/Utilities/Validator/Validator.go
@@ -1,5 +1,0 @@
-package utilities_validator
-
-type Validator interface {
-	ValidateStruct(s interface{}) error
-}
