--- Producer/UnitTest/MockItem/Repositories/OrderRepository_Impl.go
@@ -1,7 +1,0 @@
-package unittest_repositories
-
-import models "github.com/Prompiriya084/go-mq/Models"
-
-type MockOrderRepo struct {
-	*MockRepositoryImpl[models.Order]
-}
